Fix an exception when a "properties" value is not an object.
This is fixed by checking for it in the metaschema, which means it can still
happen if meta_validate is false.  I'm not sure if it should try to always
behave for any arbitrary input, even if it fails against the metaschema; I've
done it this way for now since I suspect there's a lot of broken behavior
if metaschema checks are disabled.

Note that I havn't tried to fix all possible gaps in the metaschema; for example,
additionalProperties should probably also require this.  I'm only fixing the ones
that actually cause runtime exceptions (for now).