Fix universal wheels
The current universal wheels have static install_requires matching the python version they were built on: so a wheel built on 2.6 claims it needs argparse and repoze.lru on 2.7 and 3.4, etc. This clearly doesn't work :)

environment markers are however honoured by wheels and pip, using this somewhat ugly syntax in extras (I'm looking to improve that, but this works - see the distutils-sig list for some discussion on it recently).