The new type interface: a TypeChecker class
The TypeChecker provides an immutable interface to a mapping of type
names to functions that check a value.

Includes test coverage for the new TypeChecker, including demonstrations
that it is sufficient to extend types.

The old type interface is still supported, but will raise a
DeprecationWarning. The old isinstance checks are converted to functions
& passed to the TypeChecker. The old types are also checked with the
JSON schema test suite 'type' tests.