Fix: Convert properties to string before regex
It is not uncommon for developers to create schemas in YAML, for example
when developing an Open API Specification. When keys are numeric and
parsed with the yaml parser they are typed and throw an error when
attempting to validate patternProperties or similar pattern matching
validation. This changeset converts the property to a string before
attempting to match the regex.