add parameter checking to checks()
Currently, we don't check that the arguments to checks() are actually a
string and a function (in that order). It's easy to reverse the order of
arguments, which leads to silent failure because the function is not a
match when we do a key lookup in self.checkers.

Fix this by validating the paramaters to checks().