Fix the error properties for propertyNames.
Note that propertyNames is another example (like #119)
of something that needs an additional way to say it
has errored on a *key* within the instance.

So, that's still missing, but path is not the right
thing, same as in the earlier instances.

It's also becoming somewhat clear that compound
validators should possibly grow a way to become lazy...