Specify the type of error in pretty error messages.                                                                                                                                     julian@Air ●
This still isn't perfect, there's still some refactoring to
be done to make parsing errors a bit easier to understand for
beginners.

Also there are tests missing for schema errors, both for invalid
schemas and for schemas that aren't valid JSON, both of which
should produce errors, in particular a SchemaError for the first
one, which was most of the original reason for this change (to have
users be able to differentiate between invalid instances and invalid
schemas.)

To be done in a later commit.