Move where errors are written outside of cli writers.
This makes it harder to accidentally write to different streams
from different writers.

That's tested behavior, so nothing was broken, but makes it
harder to write incorrectly in the future.

Also seems likely that something should bunch up stdin, stdout
stderr and a formatter all in one object, but that to come later.