Don't include `if` in the schema_path of errors.
Closes: #696

This is definitely not the right general solution though.

That was obvious before given that $ref needed special casing,
but it's even more obvious now.

As-is, downstream implementers of new validator keywords have
no way to get this kind of behavior. It's additional complexity
that's been introduced in modern specs, but it seems likely
that a generic mechanism now needs adding for allowing keywords
to more cleanly ignore themselves in schema paths, or otherwise
do more complex manipulation of the schema path for an error.