#708 - varying length fix
From one version of Python to another, the "type" that appears in brackets may be of differing lengths.  (E.g., `JSONDecodeError` in Python 3, but `ValueError` in Python 2.)  That would make the length of the Unicode bar different under each version of Python.  When these tests were corrected to work under Python 2, they wouldn't work under Python 3 and vice versa.  Therefore, it's simpler to break the single assertions into two shorter ones that use very little of the bar characters.