#708 - conditional message code for old Python
For Python interpreters that don't treat multibyte Unicode characters as single characters, add some code to manually calculate the number of bar characters required to reach the desired line length.  This code is confined to a conditional block and may be removed when support for older interpreters is no longer needed.