Unbreak non-format-setuptools-extra-installed jsonschema.
Revert the changes to the docs requirements.

fqdn is *not* a doc requirement, it's part of the package's
requirements, but only when format validation is installed.

The docs builds were failing because they do not install the
format extra.

This was actually a (known bad) CI gap, that we didn't run the tests
against all the configurations (i.e. against jsonschema installed with
no setuptools extras, say).

Fixed in the previous commit, so now we can run this.