Move the latest validator CLI test to where it must go now.
Checking that validator_for works correctly was not the intention of
that test.  For now, that means we need a guard to ensure we test a
validator that only would work if the latest version was used.

We can clean that up later.

Also remove test_real_validator, since there's now 3 different tests
using real validators.