Go back to fuzz_* naming.
Hypothesis isn't a current testing dependency, so as this
was before (where it was named test_* and discovered by the
test runner) there are CI failures for the normal test suite
run.

In the future hypothesis will likely become a normal testing
dependency and then we can run this as well, but for now,
only run it in the CIFuzz job.