Remove shebangs from nonexecutable scripts
When packaging jsonschema and poetry-core (which bundles it) in Fedora, we have
realised that there are nonexecutable files with a shebang line.

It seems that the primary purpose of these files is to be imported from Python code
and hence the shebangs appear to be unnecessary.

Shebangs are hard to handle when doing downstream packaging, because it makes
sense for upstream to use `#!/usr/bin/env` python while in the RPM package,
we need to avoid that and use a more specific interpreter. Since the
shebangs were unused, I propose to remove them to avoid problems.