Remove the added argument to jsonschema.validators.create.                                                                                                                julian@Airm ●
We need a better public API for vocabularies, and this unfortunately
won't cut it.

For now this is dirty, but works to load all vocabularies (which
are just concatenated together).

The reason this is needed is because I don't see a PEP302-compliant
way to load a whole directory using the importlib.resources APIs
(which call directories 'packages') -- and doing so manually would
break importing jsonschema via a zip file, which is something
historically that users were doing.