Add `jsonschema.protocols.IValidator`
This is a Protocol implementation for type checking under mypy and
other static analyzers. It uses the protocol class defined in py3.8+
and uses typing_extensions as a backport for py3.7

The documentation-only validator class has been replaced with the
protocol, and docs are now driven via autoclass on the protocol.

Importantly, several documented methods of the class have been
removed, as they were marked deprecated under jsonschema v3.0 and are
no longer provided by the builtin validators.

Minor adjustments to the docs are made to repoint references at the
new class definition.