Fix nitpick error on type annotation
Nitpick was failing on

    TYPE_CHECKER: ClassVar[TypeChecker]

It's not clear what to do about this. `TypeChecker` was imported
correctly from `jsonschema._types` and is noted in the doc as
`jsonschema.TypeChecker`. We can't import the `jsonschema` name at
runtime because that would be circular.

To resolve, use `typing.TYPE_CHECKING` to conditionally import
`jsonschema` at type-checking time. This avoids the circular
import but allows us to write

    TYPE_CHECKER: ClassVar[jsonschema.TypeChecker]

As a result, Sphinx correctly builds a cross-reference from the
annotation, the annotation is still accurate, and runtime behavior
is left untouched.