Fix calling Validator.evolve for some downstream users.
The broken case here was subclassing a validator class, something that
isn't really a supported use of the Validator classes, but of course one
can't blame anyone too much since doing so didn't raise an error or emit
any warning.

In the next release subclassing will warn, and at some point afterwards
will become an explicit error. If you're a downstream user of this
library looking for a way to avoid whichever inheritance is currently
needed in your library feel free to reach out and I'll try to help.

Closes: #982