Fix $id in the presence of $ref on draft 6 and 7.
In these drafts, not only did $ref cause other validator keywords
to be ignored, it prevents $id from setting the resolution scope
too.

In draft 3 and 4 this is now half-fixed in the sense that the
correct URI is now the one that will be retrieved, but the logic
for finding subschemas by ID is still broken on these drafts.