Update the Draft 3 and 4 metaschemas.
For draft 4 this pulls in upstream fixes which were not present locally,
notably fixing `id` to not have `format: uri` in it, because location
independent identifiers are indeed not URIs (they're URI references as
later metaschemas use).

For enum, on draft 3 and 4, this also *re-adds* constraints that enum
items MUST be unique and the array non-empty. In later drafts, this
restriction was loosened (see
json-schema-org/json-schema-spec@cf0ec7210011272f4e95f70f8c1374593e13c8c5)
as well as https://github.com/json-schema-org/json-schema-spec/issues/717#issuecomment-1299112660
but in drafts 3 and 4 it is present.

The draft 4 metaschema contains these assertions, the draft 3 one is
still buggy and does not, so they're just applied locally here.

Ref: json-schema-org/json-schema-spec#310