Move reference resolution to a (private) Validator method.
This will make it easier to swap over to referencing's
Resolver (while preserving backwards compat for anyone who
passes a RefResolver to a Validator).

At some point in the future this method may become public
(which will make it easier for external dialects to resolve
references) but let's keep it private for a bit until it's
clear that the interface is stable -- a future draft might
do crazy things like have adjacent properties to $ref affect
the resolution behavior, which would make this method need to
take more than just the $ref value.