Ignore non-str $ids for *deprecated* RefResolver resolution.
The new referencing behavior is more correct (which is why it exists),
but also means that even though `RefResolver` was/is untouched, it can
be called now with more subschemas than previously, and in some cases
that tickles this code to blow up (see the closed issue for a specific
example).

We simply now ignore non-strs, as doing so is no more wrong than this
code used to be.

Closes: #1085