Re-enable (but deprecate) automatic reference retrieval.
Changing this without deprecation is backwards incompatible, so we
re-introduce a warning.

This only applies if you have not configured neither a Registry nor a legacy
RefResolver. Both of the former 2 cases already have 'correct' behavior (the
former will not automatically retrieve references and is not backwards
incompatible as it is a new API, and the latter will do so but is already
fully deprecated by this release).

Cloess: #1089