Slightly speed up unevaluatedProperties in the failing case.
When we decide a property isn't evaluated because it fails validation
of the unevaluated* schema (which we currently do in an entirely nonideal
way anyhow by doing the work *redundantly* in two places), there's no need
to build the full list of errors when a property fails validation, it's
enough to just know it fails validation.

This may change if/when we provide full context on unevaluatedProperties
failures, but for now we don't expose that information anyhow.