Only use a default (remote retrieving registry) when none is provided.
In other words when someone does *not* do Validator(..., registry=...)

This both limits the case which we *do* still perform automatic remote
ref retrieval (in a way that still preserves backwards compatibility
given that the registry argument did not previously exist, so someone
using it must be doing something new) as well as fixes an issue with
creating a validator with a registry that has a retrieve function.

Previously the latter raised a ValueError (coming from referencing which
doesn't want you combining registries unless they agree about retrieval
functions. See the added test(s) for details).

Fixes the issue mentioned here:

https://github.com/python-jsonschema/jsonschema/issues/1065#issuecomment-1587972991

Refs: #1089