Set a library-specific user agent when automatically retrieving $refs.
This behavior is (already) deprecated, but oddly there are "real world"
webservers which seem to have allowed the `requests` default user agent
(i.e. respond correctly) but not the `urllib` default vague `Python/3.x`
one (i.e. respond with a 403 or whatever).

Since it seems polite to set this anyhow, we may as well fix behavior
for anyone who happens to be encountering such a webserver.