More closely match modern JSON Schema vocabulary in internal module names.
These are private, so it's easy to rename them to use preferred modern
terminology, namely that individual keywords are called 'keywords' rather
than 'validators'.

This leaves the latter ('validator') to have one fewer overloaded meaning
in this library, leaving it primarily referring to class objects such as
DraftNValidator objects, however there are still at least 2 public places
where we conflate terminology:

    * In the VALIDATORS attribute on validator classes, where this
      attribute is a mapping from str to callable but really the
      callables are callables for each *keyword*

    * ValidationError.validator, which is really the *keyword* which
      failed validation

These are of course public API and need deprecating, which hasn't been
done thus far mostly to not create API churn simply to rename.

In the future however, it's likely that broader deprecations will help
us deprecate these.

Specifically when we implement fuller support for vocabularies and/or
deprecate jsonschema.validators.create in favor of newer objects, we may
get a chance to replace VALIDATORS, and when we implement more robust
exception types (e.g. to address #119) we likely will deprecate
.validator.