Split the pip-licenses noxenv off to its own noxenv.
Dependencies of pip-tools shouldn't be included in the license check.

I was lazy before and knew this could be the case, but now some CI is
failing because of a new pip-tools dependency (albeit one whose license
is fine just not on the known list) but it's good to make this properly
go away as a possibility.