Don't reorder dicts (particularly schemas) when rendering errors.
Schemas, particularly longer ones, are often written intentionally with
their keywords in some specific human-understandable order.

pprint, which we currently use (for better or worse) for rendering
schemas and instances, supports *not* sorting dicts now that they
maintain their insertion order.

So we now pass that argument, and thereby preserve the order the schema
was written in.

Refs: #243