Cope with the suite permuting of remotes.
This new-but-similarly-messy logic comes modified from Bowtie
(which itself may need to be modified there, we'll see).

`jsonschema_suite remotes` is somewhat semantically broken (as it
doesn't really help with not emitting schemas which aren't valid under
the version under test), so it's time to give up and use the other
filesystem mechanism of walking the remotes directory.

Refs: json-schema-org/JSON-Schema-Test-Suite#753