Add a typing test for the Validator protocol
The typing tests are a new subdirectory of `jsonschema/tests/` which
are checked under a more stringent mypy configuration
(`--warn-unused-ignores`) in order to allow certain kinds of negative
tests against the declared types. [^1]

The first new test confirms that each validator matches the Validator
protocol, and furthermore that this is not vacuously true by way of
`Any`.

The test failed at first, as the return type of `create()` was not
annotated, and therefore under the declared types in `jsonschema`, all
of the validators were of type `Any`. To resolve, the return type of
`create` is now annotated as `type[Validator]`.

[^1]: Technically, the new nox configuration checks these files twice,
but only the second check, with `--warn-unused-ignores`, is doing all
of the necessary work.