Get CI passing again by ignoring the pip CVE.
It's not relevant to us, pip isn't a dependency.

This fix isn't quite right though as clearly we should run pip-audit in
a way where we don't get false positives when its own dependencies (here
pip) have vulnerabilities.

Running pip-audit on a library (as opposed to an end-user application)
isn't the most common use however, so it doesn't appear there's a
straightforward way to do this unless we come up with one ourselves.